mvoe to listColors
