+++ src/vs/platform/theme/common/colors/listColors.ts
@@ -7,11 +7,11 @@ import * as nls from 'vs/nls';
 
 // Import the effects we need
 import { Color } from 'vs/base/common/color';
-import { registerColor, darken, lighten, transparent, ifDefinedThenElse } from 'vs/platform/theme/common/colorUtils';
+import { darken, ifDefinedThenElse, lighten, registerColor, transparent } from 'vs/platform/theme/common/colorUtils';
 
 // Import the colors we need
-import { foreground, contrastBorder, activeContrastBorder, focusBorder, iconForeground } from 'vs/platform/theme/common/colors/baseColors';
-import { editorWidgetBackground, editorFindMatchHighlightBorder, editorFindMatchHighlight, widgetShadow } from 'vs/platform/theme/common/colors/editorColors';
+import { activeContrastBorder, contrastBorder, focusBorder, foreground, iconForeground } from 'vs/platform/theme/common/colors/baseColors';
+import { editorFindMatchHighlight, editorFindMatchHighlightBorder, editorWidgetBackground, editorWidgetForeground, widgetShadow } from 'vs/platform/theme/common/colors/editorColors';
 
 
 export const listFocusBackground = registerColor('list.focusBackground',
@@ -145,3 +145,21 @@ export const tableColumnsBorder = registerColor('tree.tableColumnsBorder',
 export const tableOddRowsBackgroundColor = registerColor('tree.tableOddRowsBackground',
 	{ dark: transparent(foreground, 0.04), light: transparent(foreground, 0.04), hcDark: null, hcLight: null },
 	nls.localize('tableOddRowsBackgroundColor', "Background color for odd table rows."));
+
+// ------ action list
+
+export const editorActionListBackground = registerColor('editorActionList.background',
+	editorWidgetBackground,
+	nls.localize('editorActionListBackground', "Action List background color."));
+
+export const editorActionListForeground = registerColor('editorActionList.foreground',
+	editorWidgetForeground,
+	nls.localize('editorActionListForeground', "Action List foreground color."));
+
+export const editorActionListFocusForeground = registerColor('editorActionList.focusForeground',
+	listActiveSelectionForeground,
+	nls.localize('editorActionListFocusForeground', "Action List foreground color for the focused item."));
+
+export const editorActionListFocusBackground = registerColor('editorActionList.focusBackground',
+	listActiveSelectionBackground,
+	nls.localize('editorActionListFocusBackground', "Action List background color for the focused item."));
